### 2025-10-30 #728 @HzaCode
Athlytics: Sports Physiology Analytics from Local Activity Files and 'Strava' Exports
Submitting Author Name: Zhiang He
Submitting Author Github Handle: @HzaCode
Repository:  https://github.com/HzaCode/Athlytics
Version submitted: 1.0.0
Submission type: Standard

Version accepted: TBD

---

* Paste the full DESCRIPTION file inside a code block below:

```
Package: Athlytics
Title: Academic R Package for Sports Physiology Analysis from Local 'Strava' Data
Version: 1.0.0
Author: Zhiang He [aut, cre]
Maintainer: Zhiang He <ang@hezhiang.com>
    person("Zhiang", "He", email = "ang@hezhiang.com", role = c("aut", "cre"))
Description: An open-source computational framework for longitudinal analysis of
    exercise physiology metrics using local 'Strava' data exports. Designed
    for personal analysis and sports science applications, this package provides
    standardized functions to calculate and visualize key physiological indicators
    including Acute:Chronic Workload Ratio (ACWR), Efficiency Factor (EF), and
    training load metrics.
    rlang (>= 0.4.0),
    viridis,
    zoo,
    R.utils,
    tools,
    utils
    devtools,
    pkgdown,
    mockery,
    XML,
    rStrava,
    xml2
    grimbough/FITfileR
NeedsCompilation: no
```

* Please indicate which category or categories from our [[package fit policies](https://devguide.ropensci.org/softwarereview_policies.html#package-categories)](https://devguide.ropensci.org/softwarereview_policies.html#package-categories) this package falls under: (Please check an appropriate box below. If you are unsure, we suggest you make a pre-submission inquiry.):

* Explain how and why the package falls under these categories (briefly, 1-2 sentences):

Athlytics provides a standardized, reproducible offline workflow for local Strava exports, covering cleaning/QC, computation of key physiology metrics (ACWR, EF, training load), and visualization. It emphasizes validation, automated pipelines, and research-grade reproducibility.

* Who is the target audience and what are scientific applications of this package?

Sports scientists, exercise physiologists, coaches, and researchers analyzing endurance training. Applications include workload monitoring (ACWR), aerobic efficiency tracking (EF), decoupling assessment, and cohort benchmarking for longitudinal studies.

* Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [[our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

`rStrava` focuses on API retrieval; `trackeR`/`activatr` provide general activity analytics. Athlytics is API-free and privacy-first, offering an end-to-end, reproducible analysis pipeline with built-in physiological models tailored for research settings.

* (If applicable) Does your package comply with our [guidance around *Ethics, Data Privacy and Human Subjects Research*](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

Yes. Analyses operate solely on user-owned local exports; no external personal data collection. Documentation encourages de-identification and minimal-data practices.

* If you made a pre-submission inquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted.

Pre-submission inquiry:
rOpenSci discussion #702 (Sports Analytics category).
JOSS submission (review paused):https://github.com/openjournals/joss-reviews/issues/9075

* Explain reasons for any [`pkgcheck` items](https://docs.ropensci.org/pkgcheck/) which your package is unable to pass.

None — all required [pkgcheck items](https://docs.ropensci.org/pkgcheck/) pass.

* [x] I have read the [[author guide](https://devguide.ropensci.org/softwarereview_author.html)](https://devguide.ropensci.org/softwarereview_author.html) and I expect to maintain this package for at least 2 years or to find a replacement.

* [x] contains a [[README with instructions for installing the development version](https://ropensci.github.io/dev_guide/building.html#readme)](https://ropensci.github.io/dev_guide/building.html#readme).
* [x] includes [[documentation with examples for all functions, created with roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation)](https://ropensci.github.io/dev_guide/building.html#documentation).
* [x] has a [[test suite](https://ropensci.github.io/dev_guide/building.html#testing)](https://ropensci.github.io/dev_guide/building.html#testing).
* [x] has [[continuous integration](https://ropensci.github.io/dev_guide/ci.html)](https://ropensci.github.io/dev_guide/ci.html), including reporting of test coverage.

* [ ] Do you wish to submit an Applications Article about your package to [[Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/%28ISSN%292041-210X/)? If so:

* [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal (see [[MEE's Policy on Publishing Code](https://besjournals.onlinelibrary.wiley.com/hub/journal/2041210x/policyonpublishingcode.html)](https://besjournals.onlinelibrary.wiley.com/hub/journal/2041210x/policyonpublishingcode.html))
* (*Scope: Do consider MEE's [[Aims and Scope](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/%28ISSN%292041-210X/aims-and-scope/read-full-aims-and-scope.html) for your manuscript. We make no guarantee that your manuscript will be within MEE scope.*)
* (*Although not required, we strongly recommend having a full manuscript prepared when you submit here.*)
* (*Please do not submit your package separately to Methods in Ecology and Evolution*)

* [x] I agree to abide by [[rOpenSci's Code of Conduct](https://ropensci.org/code-of-conduct/)](https://ropensci.org/code-of-conduct/) during the review process and in maintaining my package should it be accepted.
