### 2025-11-24 #728 @emilyriederer
Hi @HzaCode -- a few notes/questions before we proceed:

- The one item we must resolve is marked with the purple X. The package currently doesn't have [continuous integration](https://devguide.ropensci.org/pkg_ci.html)
- I see the packages imports `XML` which is obsolete. Our [recommended scaffolding](https://devguide.ropensci.org/pkg_building.html#recommended-scaffolding) suggests `xml2` as an alternative (more details at link). Have you considered migrating? Ultimately, this is your call but would be best to do before the review process
- The other really standout item on the log is the failed R CMD CHECK "no_import_package_as_a_whole". This likely means you have a library() import in one of your source files whereas the best practice is to declare dependencies and access functions as `package::function_name()`. I would recommend addressing this now since it will be a blocker from getting to CRAN
- The other standout feature is the relatively high [cyclomatic complexity](https://github.com/MangoTheCat/cyclocomp) of many functions. This may generally mean there are many "forking paths" (e.g. conditionals) within your code. I may defer this to our reviewers to help assess the necessity, but wanted to flag in case it inspires self-evident areas of improvement
