### 2025-11-29 #728 @emilyriederer
@mpadge -- FYI, do you have any idea why the "package has CICD" check would be failing again? I see it was fixed after [this comment](https://github.com/ropensci/software-review/issues/728#issuecomment-3481500222) and confirmed the README has not had reversions since
