### 2025-12-11 #728 @HzaCode
Hi @mpadge,

Thank you for looking into this and for the clarification regarding the CRAN verification logic. I appreciate the quick fix!

@emilyriederer — With this issue clarified, I am ready to proceed to the next stage.
