### 2025-12-18 #728 @HzaCode
Hi @emilyriederer,

I’ve fixed the documentation link and also addressed the non-blocking items (removed the unused dependency and reduced complexity).

Thanks again for helping find reviewers!
