### 2026-01-04 #728 @emilyriederer
One clarification for @markean and @smnnlt -- while the bot has assigned automatic due dates of 1/10 and 1/23, respectively, rOpenSci was technically "closed" for the holidays until tomorrow, Jan 5. Please feel free to consider the target date 3 weeks from that (so, 1/26) instead of the automated dates if you prefer.
