### 2026-01-05 #728 @mpadge
@emilyriederer You can get the bot to [update the due date](https://github.com/ropensci/software-review/issues/671#issuecomment-3684576946) with `set  due date for @username to YYYY-MM-DD`. Hope that helps :+1:
