### 2026-01-12 #728 @HzaCode
@markean Thank you sincerely for the thorough and insightful review. It is a pleasure to receive such constructive feedback, and I appreciate your kind words regarding the package's structure.

I fully accept your recommendations. I specifically want to acknowledge your suggestion regarding the simulated data and executable examples; implementing this will significantly enhance the package's transparency and usability for new users.

I will address all the points raised—including documentation, naming consistency, and formatting—and will ensure a proper release tag is added upon completion. Thank you for your time and guidance.
