### 2026-01-16 #728 @HzaCode
Thank you, @markean, for the thorough and constructive review.

Here is a summary of the specific changes made:

### Documentation
-   **Strava Context**: Added a brief description and link to Strava in the `README` for users unfamiliar with the platform.
-   **Package Documentation**: Added top-level `?Athlytics` documentation via `usethis::use_package_doc()`, serving as a centralized place for imports and overview.

### Code & Data
-   **Startup Message**: Removed the startup message in `zzz.R` to reduce console clutter, aligning with rOpenSci guidelines.
-   **Naming Consistency**:
    -   Adopted verb-first naming for clarity (e.g., `calculate_cohort_reference()`, `summarize_quality()`).
    -   **Dataset Naming**: Renamed sample datasets to remove the redundant prefix (e.g., `athlytics_sample_acwr` $\to$ `sample_acwr`).
-   **Reproducibility**: Added generation scripts under `data-raw/` (e.g., `generate_sample_acwr.R`) to make the simulated data fully reproducible.
-   **Runnable Examples**: Updated vignettes and Roxygen examples to use the built-in `sample_*` datasets. Key plots are now evaluated and visible directly in the documentation without requiring local Strava data files.

### Miscellaneous
-   **Formatting**: Ran `styler::style_pkg()` across the codebase to ensure consistent indentation and spacing.
-   **Versioning**: Bumped version to **v1.0.2** and created a corresponding release tag.

### References

-   **Release Tag**: [v1.0.2](https://github.com/HzaCode/Athlytics/releases/tag/v1.0.2)

---

**On a personal note:**

Although the review is still in progress, I feel compelled to share a personal thought.

It’s seriously refreshing to be here. Unlike some other environments where you have to deal with arbitrary gatekeeping or just weird politics, rOpenSci is a breath of fresh air.

It feels **fair in standards, fair in execution, and fair to new contributors**.

Appreciate the guidance from @markean, @emilyriederer, and @mpadge, and the support from the whole @ropensci community. Thanks for cutting through the noise and for creating such a wonderful community.
