### 2026-02-07 #728 @emilyriederer
Hi @HzaCode -- just to clarify: is your response above only for @smnnlt 's feedback or also @markean 's? Should they both submit their [approval template](https://devguide.ropensci.org/approval2template.html) or is there more to come wrt @markean 's comments?
