### 2026-02-20 #728 @HzaCode
Hi @emilyriederer,

Thanks so much for the approval and for guiding this process!

I've completed all the post-acceptance steps:

- ✅ **Repository transferred** to `ropensci/Athlytics`
- ✅ **All links updated** — GitHub URLs, badges (CI, Codecov, rOpenSci peer review), and documentation links now point to `ropensci/Athlytics` and `docs.ropensci.org/Athlytics`
- ✅ **`CODE_OF_CONDUCT.md` deleted** — deferring to rOpenSci's default
- ✅ **`DESCRIPTION` updated** — `URL` and `BugReports` point to the new repo; added @markean (Eunseop Kim) and @smnnlt (Simon Nolte) as `"rev"` contributors
- ✅ **pkgdown auto-deployment disabled** — removed `.github/workflows/pkgdown.yml`; `_pkgdown.yml` kept with custom styling removed, ready for rOpenSci central docs
- ✅ **R-universe install method** added to `README.md`
- ✅ **rOpenSci peer review badge** added to `README.md`

Regarding the blog post — I'd be happy to write a short technical introduction to the package. @ropensci/blog-editors please let me know when timing works or how to proceed!

Thanks again to @markean and @smnnlt for their thorough and constructive reviews!
